+++ dutchiepay/src/app/_components/_user/FindSubmit.jsx
@@ -5,7 +5,7 @@ import '@/styles/user.css';
 
 import { useEffect, useState } from 'react';
 
-import EmailInput from './EmailInput';
+import EmailInput from './_input/EmailInput';
 import Link from 'next/link';
 import PhoneAuth from './_phone/PhoneAuth';
 import axios from 'axios';
